dev: Add local docker-compose configuration
Allow running the application locally from docker compose. Since we are
going to be using a variety of services, it makes more sense to run
them in docker-compose rather than rely on people to install a bunch of
random tools on their local workstation.

This commit also adds a CI configuration to publish our Docker image to
Dockerhub.